admin_app/requirements.txt
anyio==4.3.0
apache-libcloud==3.8.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
fasteners==0.19
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy-file==0.6.0
starlette==0.37.2
starlette-admin==0.13.2
typing_extensions==4.11.0
urllib3==2.4.0
uvicorn==0.29.0
